project/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
python-multipart
sentence-transformers
faiss-cpu
numpy
pydantic[email]
